+++ src/main/assets/svgs/index.ts
@@ -2,5 +2,6 @@ import Arrow from './arrow.svg?react';
 import ArrowLeft from './arrowLeft.svg?react';
 import Lightning from './lightning.svg?react';
 import Xicon from './x-icon.svg?react';
+import Watch from './watch.svg?react';
 
-export { Arrow, ArrowLeft, Lightning, Xicon };
+export { Arrow, ArrowLeft, Lightning, Xicon, Watch };
