+++ src/main/components/AnimalProfile.tsx
@@ -1,0 +1,3 @@
+const AnimalProfile = () => {};
+
+export default AnimalProfile;
