+++ src/main/components/CustomSelect.tsx
@@ -1,4 +1,6 @@
 import { useState } from 'react';
+import { css } from '@emotion/react';
+import { theme } from '../../common/styles/theme';
 import { Arrow } from '../assets/svgs';
 
 const CustomSelect = () => {
@@ -30,3 +32,18 @@ const CustomSelect = () => {
 };
 
 export default CustomSelect;
+
+const Container = css`
+  display: flex;
+  gap: 0.5rem;
+`;
+
+const SelectBox = css`
+  width: 5.8rem;
+  color: ${theme.colors.primary};
+  appearance: none;
+  background-color: transparent;
+  border: none;
+  border-bottom: 2px solid;
+  ${theme.fonts.title};
+`;
