--- src/main/components/ProfileCard.tsx
@@ -1,3 +1,0 @@
-const ProfileCard = () => {};
-
-export default ProfileCard;
