+++ src/main/types/ProfileDataTypesProps.ts
@@ -1,0 +1,6 @@
+export interface ProfileDataTypesProps {
+  companyName: string;
+  job: string;
+  description: string;
+  nungilId: number;
+}
